feat: 근무 일정에서 평일 점심시간을 '휴가'에서 '휴게'로 변경 (#65)
- 평일 점심시간(12:00-13:00) 타이틀을 '휴가'에서 '휴게'로 변경
- 새로운 색상 매핑 추가: 휴게: "#B2DFDB"
- MyWorkComponent와 CoworkerComponent의 모든 직원 스케줄에 적용
- 드롭다운 메뉴에 '휴게' 옵션 추가
- 주말 휴가는 기존 '휴가'로 유지

변경된 파일:
- app/work/components/MyWorkComponent.jsx
- app/work/components/CoworkerComponent.jsx
