+++ lib/design-tokens.ts
@@ -38,13 +38,15 @@ export const colors = {
     },
   },
   schedule: {
-    meeting: "from-blue-50 to-blue-100 text-blue-700 border-blue-200",
-    project: "from-emerald-50 to-emerald-100 text-emerald-700 border-emerald-200",
-    education: "from-amber-50 to-amber-100 text-amber-700 border-amber-200",
-    customer: "from-purple-50 to-purple-100 text-purple-700 border-purple-200",
-    report: "from-teal-50 to-teal-100 text-teal-700 border-teal-200",
-    vacation: "from-rose-50 to-rose-100 text-rose-700 border-rose-200",
-    return: "from-indigo-50 to-indigo-100 text-indigo-700 border-indigo-200",
+    meeting: "#3b82f6", // 블루
+    project: "#10b981", // 에메랄드
+    education: "#f59e0b", // 앰버
+    customer: "#8b5cf6", // 퍼플
+    report: "#14b8a6", // 틸
+    vacation: "#f43f5e", // 로즈
+    return: "#6366f1", // 인디고
+    work: "#3b82f6", // 근무 시간 - 블루
+    break: "#22c55e", // 휴게시간 - 초록색
   },
   employee: {
     ceo: "from-purple-500 to-pink-500",
@@ -52,7 +54,7 @@ export const colors = {
     senior: "from-green-500 to-teal-500",
     designer: "from-orange-500 to-red-500",
   },
-} as const
+} as const;
 
 // 간격 (Spacing)
 export const spacing = {
@@ -63,7 +65,7 @@ export const spacing = {
   xl: "2rem", // 32px
   "2xl": "3rem", // 48px
   "3xl": "4rem", // 64px
-} as const
+} as const;
 
 // 타이포그래피
 export const typography = {
@@ -74,7 +76,7 @@ export const typography = {
   body: "text-base",
   bodySmall: "text-sm",
   caption: "text-xs",
-} as const
+} as const;
 
 // 그림자 (Shadows)
 export const shadows = {
@@ -83,14 +85,14 @@ export const shadows = {
   lg: "shadow-lg",
   xl: "shadow-xl",
   primary: "shadow-lg shadow-blue-500/25",
-} as const
+} as const;
 
 // 애니메이션 (Animations)
 export const animations = {
   transition: "transition-all duration-200",
   hover: "hover:scale-101",
   hoverUp: "hover:-translate-y-0.5",
-} as const
+} as const;
 
 // 테두리 반경 (Border Radius)
 export const borderRadius = {
@@ -98,12 +100,11 @@ export const borderRadius = {
   md: "rounded-xl",
   lg: "rounded-2xl",
   full: "rounded-full",
-} as const
+} as const;
 
 // 배경 (Backgrounds)
 export const backgrounds = {
   glass: "bg-white/60 backdrop-blur-sm",
   glassStrong: "bg-white/80 backdrop-blur-xl",
   gradient: "bg-gradient-to-br from-slate-50 via-blue-50 to-indigo-50",
-} as const
-
+} as const;
